chore: jsx -> tsx 변경
