+++ pages/point/index.tsx
